fix(badge)!: remove BaseBadge
Closes #2619
