+++ .changeset/remove-basebadge.md
@@ -1,0 +1,4 @@
+---
+"@patternfly/elements": major
+---
+`<pf-badge>`: Removed `BaseBadge` class. Reimplement (recommended) or extend `PfBadge`.
