--- elements/pf-badge/BaseBadge.css
@@ -1,6 +1,0 @@
-:host {
-  position: relative;
-  white-space: nowrap;
-  text-align: center;
-  display: inline-block;
-}
