+++ elements/pf-badge/pf-badge.ts
@@ -1,7 +1,6 @@
+import { LitElement, html } from 'lit';
 import { customElement } from 'lit/decorators/custom-element.js';
 import { property } from 'lit/decorators/property.js';
-
-import { BaseBadge } from './BaseBadge.js';
 
 import styles from './pf-badge.css';
 
@@ -31,8 +30,8 @@ import styles from './pf-badge.css';
 
 
 @customElement('pf-badge')
-export class PfBadge extends BaseBadge {
-  static readonly styles = [...BaseBadge.styles, styles];
+export class PfBadge extends LitElement {
+  static readonly styles = [styles];
 
   /**
    * Denotes the state-of-affairs this badge represents
@@ -40,9 +39,30 @@ export class PfBadge extends BaseBadge {
    */
   @property({ reflect: true }) state?: 'unread' | 'read';
 
+  /**
+   * Sets a numeric value for a badge.
+   *
+   * You can pair it with `threshold` attribute to add a `+` sign
+   * if the number exceeds the threshold value.
+   */
   @property({ reflect: true, type: Number }) number?: number;
 
+  /**
+   * Sets a threshold for the numeric value and adds `+` sign if
+   * the numeric value exceeds the threshold value.
+   */
   @property({ reflect: true, type: Number }) threshold?: number;
+
+  override render() {
+    const { threshold, number, textContent } = this;
+    const displayText =
+        (threshold && number && (threshold < number)) ? `${threshold.toString()}+`
+      : (number != null) ? number.toString()
+      : textContent ?? '';
+    return html`
+      <span>${displayText}</span>
+    `;
+  }
 }
 
 declare global {
